Renamed "DowloadDateRange"
